Remove linter errors.
